# api/openapi.yaml
openapi: 3.1.0
info:
  title: TeamTaler API
  version: 0.5.0
  description: |
    Same-origin JSON API for TeamTaler. Monetary values use integer minor units:
    response fields are exact decimal strings, while command inputs are bounded
    JSON integers.
    State-changing authenticated requests require the `X-CSRF-Token` header.
    Product creation, invitation CSV imports, and financial mutations that
    create or reverse bookings, payments, or periods additionally require an
    `Idempotency-Key`.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Create a local session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCommand'
      responses:
        '200':
          description: Authenticated session
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the current session
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Session revoked }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /session:
    get:
      tags: [Authentication]
      summary: Read the current session and group memberships
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401': { $ref: '#/components/responses/Problem' }
  /me:
    get:
      tags: [Authentication]
      summary: Read the current session and group memberships
      description: Compatibility alias for `/session`.
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401': { $ref: '#/components/responses/Problem' }
  /me/avatar:
    post:
      tags: [Authentication]
      summary: Validate, normalize, and set the current user's profile image
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Protected content-addressed profile-image URL
          content:
            application/json:
              schema:
                type: object
                required: [avatarUrl]
                properties:
                  avatarUrl: { type: string }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Authentication]
      summary: Remove the current user's profile image
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Profile image removed }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /users/{userId}/avatar/{imageKey}:
    get:
      tags: [Authentication]
      summary: Read the current profile image of a user sharing a group
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: imageKey
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-f]{64}\.png$'
      responses:
        '200':
          description: Normalized profile image
          content:
            image/png:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /invitations/preview:
    post:
      tags: [Authentication]
      summary: Resolve safe onboarding hints for a valid invitation token
      description: Rate-limited public lookup that never exposes the invited email address, roles, or category grants.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvitationPreviewCommand' }
      responses:
        '200':
          description: Safe invitation preview
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationPreview' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /invitations/accept:
    post:
      tags: [Authentication]
      summary: Consume an invitation and create or validate an account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationAcceptance'
      responses:
        '201':
          description: Joined group and authenticated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /groups:
    get:
      tags: [Groups]
      summary: List groups available to the current user
      responses:
        '200':
          description: Active group memberships
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Group' }
    post:
      tags: [Groups]
      summary: Create a group owned by the current user
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, currency]
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                currency: { type: string, pattern: '^[A-Z]{3}$' }
      responses:
        '201':
          description: Created group
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        '422': { $ref: '#/components/responses/Problem' }
  /groups/{groupId}:
    patch:
      tags: [Groups]
      summary: Update administrator-managed group settings
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
      responses:
        '200':
          description: Persisted normalized group settings
          content:
            application/json:
              schema:
                type: object
                required: [name]
                properties:
                  name: { type: string }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /groups/{groupId}/settings:
    get:
      tags: [Groups]
      summary: Read administrator-managed group behavior settings
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Current typed group settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupSettings' }
        '403': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Groups]
      summary: Update administrator-managed group behavior settings
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupSettingsUpdate' }
      responses:
        '200':
          description: Persisted typed group settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupSettings' }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /groups/{groupId}/dashboard:
    get:
      tags: [Groups]
      summary: Read the active member dashboard
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Dashboard read model
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dashboard' }
        '403': { $ref: '#/components/responses/Problem' }
  /groups/{groupId}/logo:
    post:
      tags: [Groups]
      summary: Validate, normalize, and set the group logo
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Normalized group-logo URL
          content:
            application/json:
              schema:
                type: object
                required: [logoUrl]
                properties:
                  logoUrl: { type: string }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Groups]
      summary: Restore the default TeamTaler logo for the group
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Custom group logo removed }
        '403': { $ref: '#/components/responses/Problem' }
  /groups/{groupId}/members:
    get:
      tags: [Groups]
      summary: List group members and effective permissions
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Group members
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Membership' }
  /groups/{groupId}/members/{membershipId}/permissions:
    patch:
      tags: [Groups]
      summary: Replace group roles and category grants
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/MembershipId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PermissionUpdate' }
      responses:
        '204': { description: Permissions replaced }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /groups/{groupId}/members/{membershipId}:
    delete:
      tags: [Groups]
      summary: Archive a membership while preserving all historical records
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/MembershipId'
        - $ref: '#/components/parameters/CSRFToken'
        - name: confirmSelf
          in: query
          required: false
          schema: { type: boolean, default: false }
          description: Must be true when an administrator removes their own membership.
      responses:
        '204': { description: Membership archived and effective permissions cleared }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /groups/{groupId}/accounts:
    get:
      tags: [Finance]
      summary: List consolidated member balances as a finance manager
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Active and archived group membership balances
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AccountSummary' }
        '403': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: teamtaler_session
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 1 }
...
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    LoginCommand:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, format: password, minLength: 12, maxLength: 1024 }
...
    InvitationPreviewCommand:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 1 }
...
    Session:
      type: object
      required: [user, csrfToken, groups]
      properties:
        user: { $ref: '#/components/schemas/User' }
        csrfToken: { type: string }
        groups:
          type: array
          items: { $ref: '#/components/schemas/Group' }
        activeGroupId:
          type: [string, 'null']
...